Implement Search logic - Create search component - Implement MUI AutoComplete - Fix responsive in top bar component
