+++ src/common/types/top-bar/index.ts
@@ -1,4 +1,5 @@
 export interface ITopBarProps {
   isOpen: boolean
   setIsOpen: (value: boolean) => void
+  isNonMobile: boolean
 }
