+++ src/components/search-bar/index.tsx
@@ -1,0 +1,1 @@
+export { SearchBarComponent } from "./SearchBarComponent";
