+++ src/components/theme-switcher/index.tsx
@@ -1,0 +1,3 @@
+export { ThemeSwitcherComponent } from "./ThemeSwitcherComponent";
+
+
