+++ src/components/theme-switcher/style.module.css
@@ -1,0 +1,8 @@
+
+.iconBlock {
+  padding-right: 35px;
+  padding-top: 10px;
+}
+.themeIcon {
+  margin-right: 20px !important;
+}
